requirements.txt
discord.py
ffmpeg
youtube_dl
PYNaCl
gtts
pydub